[HipFlex] First draft of the calibration entity for the hip flexibility
Take in entry the current torque and the desired position of the joints.
Output the angular correction to apply on the hip (theta_diffSOUT) and the desired positions updated with these angles (q_cmdSOUT).
Done on left and right hips. 
Possibility to activate angular saturation and lowPass filter.
Test uses the entity with a constant desired position which is the halfsitting, with same left and right hip flexibility.